Changes baseSalary type to BigDecimal
Updates the data type of the baseSalary field from Double to BigDecimal across the application.
This change ensures greater precision when handling monetary values, preventing potential issues with floating-point arithmetic.

+++ domain/model/src/main/java/co/com/pragma/model/user/User.java
@@ -6,6 +6,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -20,7 +21,7 @@ public class User {
     private String password;
     private String idNumber;
     private Long phone;
-    private Double baseSalary;
+    private BigDecimal baseSalary;
     private Role role;
     private LocalDate birthDate;
     private String address;
