+++ infrastructure/driven-adapters/r2dbc-postgresql/src/main/java/co/com/pragma/r2dbc/entity/UserEntity.java
@@ -9,6 +9,7 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.Column;
 import org.springframework.data.relational.core.mapping.Table;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.time.LocalDate;
 
@@ -28,7 +29,7 @@ public class UserEntity {
     private String password;
     private String idNumber;
     private Long phone;
-    private Double baseSalary;
+    private BigDecimal baseSalary;
     private LocalDate birthDate;
     private String address;
